Add basic DUB plugin for snapcraft
This plugin provides the essential basics to allow a DUB project to be
built; more specifically, to allow DUB itself to be built!  The defined
plugin-specific keywords allow for some small customization:

  * `build-flags`: list of flags to pass to the `dub build` command
    (e.g. `--build=release`, `--compiler=ldc2`, etc.);

  * `build-target`: the DUB target to build from `dub.json`;

  * `target-path`: the path where DUB will place built files,
    relative to the root of the project directory.

The last of these is probably the most important.  Since DUB lacks any
`install` option, the snapcraft build stage would by default complete
without any files in the `parts/[dub-project]/install/` directory.  In
order to avoid this, the plugin will by default copy the entire contents
of the corresponding `parts/[dub-project]/build/` directory, which will
include not only built files, but the entire project source.

If however `target-path` is specified, then files will be copied only
from `parts/[dub-project]/build/[target-path]`, which in most cases
ought to correspond to only the built files (assuming of course that the
the project being built has a corresponding `targetPath` defined in its
`dub.json`).