Add config files required to build and use LDC internally
As a temporary measure, in the absence of a `d-compiler` plug, the DUB
snap will build its own D compiler for internal use.  This copies the
corresponding setup for the ldc2 external snap, which in turn requires
GCC to be available for the final linking stage.

Note that this in-snap compiler will suffer from the same known issues
as the ldc2 snap, namely the inability to link against system libraries
not already included in the snap itself.