Add snapcraft.yaml defining the snap package and its components
The dub package uses `classic` confinement to ensure that it can access
the resources of the host system it is installed on, e.g. any installed
D compilers, libraries required for linking, and so on.

Note that a `network` plug is not required, since `classic` confinement
already grants this access.