Add snapcraft.yaml defining the snap package and its components
As a temporary measure, in the absence of a `d-compiler` plug, the DUB
snap will include its own D compiler for internal use.  This copies the
corresponding setup for the ldc2 external snap, which in turn requires
GCC to be available for the final linking stage.

Note that this in-snap compiler will suffer from the same known issues
as the ldc2 snap, namely the inability to link against system libraries
not already included in the snap itself.

The snap also includes a copy of curl in order to allow dub to download
packages from code.dlang.org.